"Changed carousel class to carousel-slide and added previous button for carousel"
The diff shows changes made to the home.component.html file in the home component of the app/components/pages/home directory. The changes include:

* Changing the class "carousel slide" to "carousel-slide" for the carousel div.
* Adding a new button for the previous slide with the (click) event handler "getPrev()".

Note: The diff also shows the removal of a previous button outside the carousel div, but this seems to be a mistake as the same button is added again inside the carousel div.